fix: refac code + set default bg to black
